# Make bulk loaded rtrees fully balanced